backend/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.32
ollama>=0.3.3
httpx>=0.27
beautifulsoup4>=4.12
aiosqlite>=0.19
python-multipart>=0.0.9
pydantic>=2.10
pydantic-settings>=2.6
aiosmtplib==3.0.2
dnspython==2.6.1
validators==0.22.0
apify-client>=1.6
python-dotenv>=1.0
openpyxl>=3.1
PyJWT>=2.8
httpx>=0.27
pyotp>=2.9
